ELI-136: Now able to delete Contact Groups with relations, this deletes the associated relations as well. A new prompt is added which warns the user of the implications of deleting the Contact Group. Removed the disableDelete() related tests, and added a new test for the new delete mechanic

+++ backend/dom/src/main/java/org/incode/eurocommercial/contactapp/dom/group/ContactGroup.java
@@ -214,15 +214,20 @@ public class ContactGroup extends ContactableEntity implements Comparable<Contac
 
     //region > delete (action)
 
-    @Action(semantics = SemanticsOf.IDEMPOTENT_ARE_YOU_SURE)
+    @Action(semantics = SemanticsOf.IDEMPOTENT)
     @ActionLayout(named = "Delete", position = ActionLayout.Position.PANEL)
-    public void delete() {
-        contactGroupRepository.delete(this);
+    public void delete(final @ParameterLayout(named = "This might also delete all Contact Roles and Contacts connected to it, do you wish to proceed?") boolean delete) {
+        if (delete)
+            contactGroupRepository.delete(this);
     }
 
-    public String disableDelete() {
-        return getContactRoles().isEmpty() ? null : "This group has contacts";
+    public String validateDelete(final boolean delete) {
+        return delete ? null : "You have to agree";
     }
+
+//    public String disableDelete() {
+//        return getContactRoles().isEmpty() ? null : "This group has contacts";
+//    }
 
     @Persistent(mappedBy = "contactGroup", dependentElement = "true")
     @Collection()
